Add required packages to setup.py for pip installation of dependencies.
The required dependencies can be automatically installed via pip. Numpy,
scipy and matplotlib can be installed cleanly as wheel packages. However,
pyaudio requires python.h and portaudio.h to be present since pyaudio is
compiled on the fly. Therefore it is set as an optional install. A
warning is printed when importing if pyaudio isn't installed, so that
the zignal library can be used anyway.