Add property getters and setters for f0, gain and Q values.
Removed the internal verification method and replaced it with
property decorators instead, hiding the variables as private.
Also added an internal update() method that is called every time
the setters are used.

This makes it easier to create an instance of a parametric
biquad class and update any of the values individually.