Remove nose as unit test runner
Minor bugfixes to the tests to make them run again.

Comment out direct prints in the unit tests. Nosetest used to hide
prints to stdout but pythons builtin unittest framework does not.
So a quick fix for now is to just comment out the printouts. Might
update later to use the logging feature better, if needed.

Fixes #8
Closes #9